requirements.txt
django==1.6
django-security==1.0
django-gaekit==0.2.7
